Move high-level API stuff into `Server` module
_Again_
