+++ drama.cabal
@@ -63,6 +63,8 @@ library
     , Drama.Process
     , Drama.Process.Internal
     , Drama.Reexports
+    , Drama.Server
+    , Drama.Server.Internal
 
 
 benchmark bench
