+++ src/Drama/Reexports.hs
@@ -4,9 +4,6 @@
 -- License:    BSD-3-Clause
 -- Copyright:  © 2021 Evan Relf
 -- Maintainer: evan@evanrelf.com
---
--- TODO
---
 
 module Drama.Reexports (module X) where
